+++ src/main/java/shadows/placebo/json/PSerializer.java
@@ -1,16 +1,11 @@
 package shadows.placebo.json;
 
-import java.lang.invoke.MethodHandle;
-import java.lang.invoke.MethodHandles;
-import java.lang.reflect.Method;
-import java.lang.reflect.Modifier;
-import java.util.function.BiFunction;
+import java.util.Objects;
 import java.util.function.Consumer;
 import java.util.function.Supplier;
 
-import com.google.common.base.Preconditions;
-import com.google.common.collect.BiMap;
-import com.google.gson.JsonElement;
+import javax.annotation.Nullable;
+
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParseException;
 import com.mojang.serialization.Codec;
@@ -19,32 +14,53 @@ import com.mojang.serialization.JsonOps;
 import net.minecraft.nbt.CompoundTag;
 import net.minecraft.nbt.NbtOps;
 import net.minecraft.network.FriendlyByteBuf;
-import net.minecraft.resources.ResourceLocation;
 import shadows.placebo.Placebo;
 import shadows.placebo.json.JsonUtil.JsonDeserializer;
 import shadows.placebo.json.JsonUtil.JsonSerializer;
 import shadows.placebo.json.JsonUtil.NetDeserializer;
 import shadows.placebo.json.JsonUtil.NetSerializer;
-import shadows.placebo.json.PlaceboJsonReloadListener.TypeKeyed;
 
+/**
+ * A PSerializer is a combination of json and network de/serializer.<br>
+ * Each component is optional (as necessary) except for the Json Deserializer.<br>
+ * The consumer is responsible for verifying that all serializers conform to application-dependent requirements through {@link #validate(boolean)}
+ *
+ * @param <V> The type of object this serializer interfaces with.
+ */
 public class PSerializer<V> implements JsonDeserializer<V>, JsonSerializer<V>, NetDeserializer<V>, NetSerializer<V> {
 
 	private final String name;
+
 	private final JsonDeserializer<V> jds;
+
+	@Nullable
 	private final JsonSerializer<V> js;
+
+	@Nullable
 	private final NetDeserializer<V> nds;
+
+	@Nullable
 	private final NetSerializer<V> ns;
 
-	public PSerializer(String name, JsonDeserializer<V> jds, JsonSerializer<V> js, NetDeserializer<V> nds, NetSerializer<V> ns) {
-		this.name = name;
-		this.jds = jds;
+	/**
+	 * Creates a new PSerializer.
+	 * @param name A string name of the object being de/serialized for error logging.
+	 * @param jds The json deserializer.
+	 * @param js The json serializer.
+	 * @param nds The network deserializer.
+	 * @param ns The network serializer.'
+	 * @see PSerializer.Builder
+	 */
+	public PSerializer(String name, JsonDeserializer<V> jds, @Nullable JsonSerializer<V> js, @Nullable NetDeserializer<V> nds, @Nullable NetSerializer<V> ns) {
+		this.name = Objects.requireNonNull(name);
+		this.jds = Objects.requireNonNull(jds, "The serializer for " + this.name + " is missing a json deserializer.");
 		this.js = js;
 		this.nds = nds;
 		this.ns = ns;
 	}
 
 	@Override
-	public JsonElement write(V src) {
+	public JsonObject write(V src) {
 		if (this.js == null) throw new UnsupportedOperationException("Attempted to serialize a " + this.name + " to json, but this serializer does not support that operation.");
 		return this.js.write(src);
 	}
@@ -56,8 +72,7 @@ public class PSerializer<V> implements JsonDeserializer<V>, JsonSerializer<V>, N
 	}
 
 	@Override
-	public V read(JsonElement json) throws JsonParseException {
-		if (this.jds == null) throw new UnsupportedOperationException("Attempted to deserialize a " + this.name + " from json, but this serializer does not support that operation.");
+	public V read(JsonObject json) throws JsonParseException {
 		return this.jds.read(json);
 	}
 
@@ -68,106 +83,99 @@ public class PSerializer<V> implements JsonDeserializer<V>, JsonSerializer<V>, N
 	}
 
 	@SuppressWarnings("unchecked")
-	private static <V> BiFunction<Object, Object, V> coerce(Method m) {
-		MethodHandle k;
-		try {
-			k = MethodHandles.lookup().unreflect(m);
-		} catch (IllegalAccessException e1) {
-			throw new RuntimeException(e1);
-		}
-		return (p1, p2) -> {
-			try {
-				return (V) (p1 == null ? k.invoke(p2) : k.invoke(p1, p2));
-			} catch (Throwable e) {
-				throw new RuntimeException(e);
-			}
-		};
+	public final JsonObject writeUnchecked(Object src) {
+		return write((V) src);
+	}
+
+	@SuppressWarnings("unchecked")
+	public final void writeUnchecked(Object src, FriendlyByteBuf buf) {
+		write((V) src, buf);
 	}
 
 	/**
-	 * Automatically generates a serializer based on methods found on the class.
-	 * Searches for the following methods:
-	 *
-	 * public static V read(JsonElement) [Required]
-	 * public JsonElement write() [Optional]
-	 * 
-	 * public static V read(FriendlyByteBuf) [Required if synced]
-	 * public void write(FriendlyByteBuf) [Required if synced]
-	 * 
-	 * Note: the json methods work if the json param is any subclass of JsonElement, but it must match.
-	 * 
-	 * @param name Name of the object being serialized, for error logging
-	 * @param synced If this serialized will be used for networked data
-	 * @param clazz The class containing the read/write methods
-	 * 
-	 * @deprecated Prefer using {@link #fromCodec(String, Codec)} as it gives better error reporting even when buried in lambda soup.
+	 * Validates optional properties of this PSerializer.
+	 * @param serializable If true, the Json Serializer is required.
+	 * @param synced If true, both a Network Deserializer and Network Serializer are required.
 	 */
-	@Deprecated(forRemoval = true)
-	public static <V> PSerializer.Builder<V> autoRegister(String name, Class<? extends V> clazz) {
-		Builder<V> builder = new Builder<>(name);
-		Method[] methods = clazz.getDeclaredMethods();
-		Class<?> jsonType = null;
-		for (Method m : methods) {
-			if (Modifier.isStatic(m.getModifiers())) {
-				if (m.getName().equals("read")) {
-					Class<?>[] p = m.getParameterTypes();
-					if (p.length == 1) {
-						if (p[0] == FriendlyByteBuf.class) {
-							builder.withNetworkDeserializer(buf -> PSerializer.<V>coerce(m).apply(null, buf));
-						} else if (JsonElement.class.isAssignableFrom(p[0])) {
-							builder.withJsonDeserializer(obj -> PSerializer.<V>coerce(m).apply(null, obj));
-							if (jsonType == null) jsonType = p[0];
-							else if (jsonType != p[0]) throw new RuntimeException("Invalid automatic PSerializer registration - JSON read/write type mismatch!");
-						}
-					}
-				}
-			} else {
-				if (m.getName().equals("write")) {
-					Class<?>[] p = m.getParameterTypes();
-					if (p.length == 1) {
-						if (p[0] == FriendlyByteBuf.class) {
-							builder.withNetworkSerializer((inst, buf) -> coerce(m).apply(inst, buf));
-						}
-					} else if (p.length == 0 && JsonElement.class.isAssignableFrom(m.getReturnType())) {
-						builder.withJsonSerializer((inst) -> PSerializer.<JsonElement>coerce(m).apply(null, inst));
-						if (jsonType == null) jsonType = m.getReturnType();
-						else if (jsonType != m.getReturnType()) throw new RuntimeException("Invalid automatic PSerializer registration - JSON read/write type mismatch!");
-					}
-				}
-			}
+	public void validate(boolean serializable, boolean synced) {
+		if (serializable) {
+			Objects.requireNonNull(this.js, "The serializer for " + this.name + " is missing a json serializer.");
 		}
-
-		return builder;
+		if (synced) {
+			Objects.requireNonNull(this.nds, "The serializer for " + this.name + " is missing a network deserializer.");
+			Objects.requireNonNull(this.ns, "The serializer for " + this.name + " is missing a network serializer.");
+		}
 	}
 
-	@SuppressWarnings({ "unchecked", "rawtypes" })
-	public static <V> PSerializer.Builder<V> fromCodec(String name, Codec<? extends V> codec) {
+	/**
+	 * Automatically creates a PSerializer from a codec.<br>
+	 * Network de/serialization is handled by converting the object to/from NBT and sending that over the network.
+	 * @param <V> The type of object being serialized.
+	 * @param name The name of the type of object being serialized, for error logging.
+	 * @param codec The codec.
+	 * @return A Codec-Backed PSerializer.
+	 */
+	public static <V> PSerializer<V> fromCodec(String name, Codec<V> codec) {
 		Builder<V> builder = new Builder<>(name);
-		Codec<V> rawCodec = (Codec) codec;
 		Consumer<String> onErr = msg -> logCodecError(name, msg);
-		builder.withJsonSerializer(obj -> rawCodec.encodeStart(JsonOps.INSTANCE, obj).getOrThrow(false, onErr));
-		builder.withJsonDeserializer(json -> rawCodec.decode(JsonOps.INSTANCE, json).getOrThrow(false, onErr).getFirst());
-		builder.withNetworkSerializer((obj, buf) -> buf.writeNbt((CompoundTag) rawCodec.encodeStart(NbtOps.INSTANCE, obj).getOrThrow(false, onErr)));
-		builder.withNetworkDeserializer(buf -> rawCodec.decode(NbtOps.INSTANCE, buf.readNbt()).getOrThrow(false, onErr).getFirst());
-		return builder;
+		builder.toJson(obj -> codec.encodeStart(JsonOps.INSTANCE, obj).getOrThrow(false, onErr).getAsJsonObject());
+		builder.fromJson(json -> codec.decode(JsonOps.INSTANCE, json).getOrThrow(false, onErr).getFirst());
+		builder.toNetwork((obj, buf) -> buf.writeNbt((CompoundTag) codec.encodeStart(NbtOps.INSTANCE, obj).getOrThrow(false, onErr)));
+		builder.fromNetwork(buf -> codec.decode(NbtOps.INSTANCE, buf.readNbt()).getOrThrow(false, onErr).getFirst());
+		return builder.build();
 	}
 
 	private static void logCodecError(String name, String msg) {
 		Placebo.LOGGER.error("Codec failure for type {}, message: {}", name, msg);
 	}
 
-	public static <V> PSerializer.Builder<V> builtin(String name, Supplier<V> factory) {
+	/**
+	 * Creates a PSerializer for a built-in object that is not truly loaded from json.
+	 * @param <V> The type of the object.
+	 * @param name The name of the type of the object.
+	 * @param factory A supplier to the built-in object.
+	 * @return A Serializer that will always return the built-in object.
+	 */
+	public static <V> PSerializer<V> builtin(String name, Supplier<V> factory) {
 		Builder<V> builder = new Builder<>(name);
-		builder.withJsonDeserializer(json -> factory.get()).withJsonSerializer(o -> new JsonObject());
-		builder.withNetworkDeserializer(net -> factory.get()).withNetworkSerializer((obj, buf) -> {
+		builder.fromJson(json -> factory.get()).toJson(o -> new JsonObject());
+		builder.fromNetwork(net -> factory.get()).toNetwork((obj, buf) -> {
 		});
-		return builder;
+		return builder.build();
 	}
 
-	public static <V extends TypeKeyed<V>> PSerializerTypeAdapter<V> makeTypeAdapter(BiMap<ResourceLocation, PSerializer<V>> serializers) {
-		return new PSerializerTypeAdapter<>(serializers);
+	/**
+	 * Creates a simple PSerializer which only supports json deserialization.
+	 * @param <V> The type of the object.
+	 * @param name The name of the type of the object.
+	 * @param jds The Json Deserializer for the object.
+	 * @return A Serializer with only the ability to deserialize json.
+	 */
+	public static <V> PSerializer<V> basic(String name, JsonDeserializer<V> jds) {
+		return builder(name, jds).build();
 	}
 
+	/**
+	 * Creates a new builder.
+	 * @see Builder#Builder(String)
+	 */
+	public static <V> PSerializer.Builder<V> builder(String name) {
+		return new PSerializer.Builder<>(name);
+	}
+
+	/**
+	 * Creates a new builder.
+	 * @see Builder#Builder(String)
+	 */
+	public static <V> PSerializer.Builder<V> builder(String name, JsonDeserializer<V> jds) {
+		return new PSerializer.Builder<V>(name).fromJson(jds);
+	}
+
+	/**
+	 * A PSerializer Builder is a helper class for constructing a PSerializer.
+	 *
+	 * @param <V> The type being de/serialized.
+	 */
 	public static class Builder<V> {
 
 		private final String name;
@@ -184,43 +192,77 @@ public class PSerializer<V> implements JsonDeserializer<V>, JsonSerializer<V>, N
 			this.name = name;
 		}
 
-		public Builder<V> withJsonDeserializer(JsonDeserializer<V> jds) {
+		/**
+		 * Sets the Json Deserializer of this builder.
+		 * @param jds The Json Deserializer.
+		 * @return this
+		 */
+		public Builder<V> fromJson(JsonDeserializer<V> jds) {
 			this.jds = jds;
 			return this;
 		}
 
-		public Builder<V> withJsonSerializer(JsonSerializer<V> js) {
+		/**
+		 * Sets the Json Serializer of this builder.
+		 * @param js The Json Serializer.
+		 * @return this
+		 */
+		public Builder<V> toJson(JsonSerializer<V> js) {
 			this.js = js;
 			return this;
 		}
 
-		public Builder<V> withNetworkDeserializer(NetDeserializer<V> nds) {
+		/**
+		 * Sets the Network Deserializer of this builder.
+		 * @param nds The Network Deserializer.
+		 * @return this
+		 */
+		public Builder<V> fromNetwork(NetDeserializer<V> nds) {
 			this.nds = nds;
 			return this;
 		}
 
-		public Builder<V> withNetworkSerializer(NetSerializer<V> ns) {
+		/**
+		 * Sets the Network Serializer of this builder.
+		 * @param ns The Network Serializer.
+		 * @return this
+		 */
+		public Builder<V> toNetwork(NetSerializer<V> ns) {
 			this.ns = ns;
 			return this;
 		}
 
-		public Builder<V> json(JsonDeserializer<V> jds, JsonSerializer<V> js) {
-			return this.withJsonDeserializer(jds).withJsonSerializer(js);
+		/**
+		 * Sets the Network Serializer and Deserializer of this builder.
+		 * @param nds The Network Deserializer.
+		 * @param ns The Network Serializer.
+		 * @return this
+		 */
+		public Builder<V> networked(NetDeserializer<V> nds, NetSerializer<V> ns) {
+			return this.fromNetwork(nds).toNetwork(ns);
 		}
 
-		public Builder<V> net(NetDeserializer<V> jds, NetSerializer<V> js) {
-			return this.withNetworkDeserializer(jds).withNetworkSerializer(js);
-		}
-
-		public PSerializer<V> build(boolean synced) {
-			Preconditions.checkNotNull(this.jds, "Attempted to build a Serializer for " + this.name + " but no json deserializer was provided.");
-			if (synced) {
-				Preconditions.checkNotNull(this.nds, "Attempted to build a Synced Serializer for " + this.name + " but no network deserializer was provided.");
-				Preconditions.checkNotNull(this.ns, "Attempted to build a Synced Serializer for " + this.name + " but no network serializer was provided.");
-			}
+		/**
+		 * Builds the underlying PSerializer.
+		 */
+		public PSerializer<V> build() {
 			return new PSerializer<>(this.name, this.jds, this.js, this.nds, this.ns);
 		}
 
+	}
+
+	/**
+	 * An interface supporting Serializers with generic types.
+	 *
+	 * @param <V> This
+	 */
+	public static interface PSerializable<V extends PSerializable<V>> {
+
+		/**
+		 * Returns the serializer that is responsible for de/serializing this object.<br>
+		 * If this object is subtyped (i.e. is backed by a {@link SerializerMap}, the returned serializer should be registered.
+		 */
+		PSerializer<? extends V> getSerializer();
 	}
 
 }